requirements.txt
numpy>=1.26
opencv-python>=4.10
Pillow>=10.0
torch>=2.3
torchvision>=0.18
